added UV Index apis and docs
