+++ README.md
@@ -16,6 +16,36 @@ Currently only the current weather, 5day, and 16day forecast apis have been adde
 npm install openweathermap-apis
 ```
 
+## usage
+
+To use this package you can either import an individual api or import all of the apis into your application. For an individual module you can use:
+
+```
+import Weather from 'OpenWeatherMap/api';
+```
+or 
+```
+const { Weather } = require('openweathermap-api);
+```
+
+If you want to import all of the modules just use:
+
+```
+const OWM = require('openweathermap-api');
+```
+
+Once imported you need to create a client instance where wou would provide an api key for it to use:
+
+```
+const { Weather } = require('openweathermap-api');
+
+const client = new Weather({
+  apiKey: '{yourAPPIDFromopenweathermap.org}'
+});
+```
+
+If you have imported all of the apis into a single object then you will have full access to all methods as if you had imported only a single one.
+
 ## api documentation
 
 The documentation can be found on the [API Docs Page](./docs/api.md)
