+++ index.js
@@ -1,8 +1,197 @@
+'use strict';
+
+const request = require('request-promise-native');
+
+const WeatherMixin = require('./lib/weather');
+const UVIndexMixin = require('./lib/uvIndex');
+
 /**
- * @module OpenWeatherMap
- * @description An abstraction layer for accessing the various OpenWeatherMap APIs.
+ * @module OpenWeatherMap/api
+ * @description The OpenWeatherMap/api module acts as an abstraction layer for accessing the various OpenWeatherMap APIs.
  * @see {@link https://openweathermap.org/api}
- * @exports Weather
+ * @exports external:Weather
  */
 
-module.exports.Weather = require('./lib/weather');
+/**
+ * @class
+ * @memberof module:OpenWeatherMap/api
+ */
+class OpenWeatherMap {
+  /**
+   * @constructor
+   * @param {Object} options
+   * @param {String} options.apiKey the APPID supplied by openweathermap.org
+   * @param {String} [options.language] an optional 2 letter language code e.g. 'en'
+   * @param {String} [options.hostname=api.openweathermap.org] an optional hostname for the api to connect to.
+   * @param {String|Number} [options.port=80] an optional port to use for the api
+   */
+  constructor(options) {
+    this.apiKey = options.apiKey || '';
+    this.language = options.language || undefined;
+    this.hostname = options.hostname || 'api.openweathermap.org';
+    this.port = `${options.port}` || '80';
+  }
+
+  /**
+   * @method module:OpenWeatherMap/api.OpenWeatherMap~sendRequest
+   * @desc builds and sends the request to the OpenWeatherMap API
+   * @param {String} uri the uri to send the request too
+   * @param {Object} params
+   * @returns {Promise} the http request promise
+   */
+  sendRequest(uri, params) {
+    const options = {
+      uri: `http://${this.hostname}${uri}`,
+      qs: Object.assign(
+        {
+          APPID: `${this.apiKey}`,
+          units: this.units,
+          lang: this.language
+        },
+        parseParameters(params)
+      ),
+      json: true
+    };
+    return request(options);
+  }
+}
+
+/**
+ * @method mix
+ * @desc mixes a superclass using the MixinBuilder
+ * @param {Class} Superclass 
+ * @returns {MixinBuilder} returns a MixinBuilder ready to apply mixins
+ * @private
+ */
+const mix = (Superclass) => {
+  return new MixinBuilder(Superclass);
+};
+
+/**
+ * @class
+ * @classdesc a helper class for constructing a class with mixins
+ * @private
+ */
+class MixinBuilder {
+  /**
+   * @constructor
+   * @param {Class} Superclass the parent class to have the mixins applied too
+   * @private
+   */
+  constructor(Superclass) {
+    this.superclass = Superclass;
+  }
+  
+  /**
+   * @method MixinBuilder~with
+   * @desc specifies whch mixins to apply to the super class
+   * @param {...Function} mixins 
+   * @returns {Class} returns the superclass with the mixins applied
+   * @private
+   */
+  with(...mixins) {
+    return mixins.reduce((c, mixin) => { return mixin(c); }, this.superclass);
+  }
+}
+
+/**
+ * @method parseParameters
+ * @instance
+ * @desc parses the modules parameters into the api parameters
+ * @param {RequestParameters} params
+ * @returns {Object} returns a query object with the correct formatting when sending the search query
+ * @private
+ */
+function parseParameters(params) {
+  const finalParams = {};
+
+  // by city ID (prefered)
+  if (params.id && params.id !== '') {
+    finalParams.id = `${params.id}`;
+
+  // by city name and country
+  } else if (params.city && params.city !== ''
+      && params.country && params.country !== '') {
+    finalParams.q = `${params.city},${params.country}`;
+
+  // by just city name
+  } else if (params.city && params.city !== '') {
+    finalParams.q = `${params.city}`;
+
+  // by lat long
+  } else if (params.coordinates
+      && params.coordinates.latitude && params.coordinates.latitude !== ''
+      && params.coordinates.longitude && params.coordinates.longitude !== '') {
+    finalParams.lat = `${params.coordinates.latitude}`;
+    finalParams.lon = `${params.coordinates.longitude}`;
+
+  // by zip and country
+  } else if (((params.zip && params.zip !== '') || (params.postcode && params.postcode !== ''))
+      && params.country && params.country !== '') {
+    finalParams.zip = `${params.zip || params.postcode},${params.country}`;
+
+  // just by zip code (USA only)
+  } else if ((params.zip && params.zip !== '') || (params.postcode && params.postcode !== '')) {
+    finalParams.zip = `${params.zip || params.postcode}`;
+  }
+
+  // an optional days limit (only used on dailyForecast)
+  if ((params.days && params.days !== '') || (params.hours && params.hours !== '')) {
+    finalParams.cnt = `${params.days || params.hours}`;
+  }
+
+  return finalParams;
+}
+
+
+module.exports = mix(OpenWeatherMap).with(WeatherMixin, UVIndexMixin);
+module.exports.Weather = mix(OpenWeatherMap).with(WeatherMixin);
+module.exports.UVIndex = mix(OpenWeatherMap).with(UVIndexMixin);
+
+/**
+ * @typedef {Object} CityIDReqParams
+ * @global
+ * @description an object of parameters used to identify a location using a citys ID
+ * @property {String|Number} id the city ID to be used in the query
+ * @property {String|Number} [days] number of days to return in the forecast (used only in dailyForecast alias for cnt)
+ * @property {String|Number} [hours] number of hours to return in the forecast (used only in forecast alias for cnt)
+ */
+
+/**
+ * @typedef {Object} CityNameReqParams
+ * @global
+ * @description an object of parameters used to identify a location using a citys name
+ * @property {String} city the city name to use in the query e.g. 'london' (best used with country also set)
+ * @property {String} [country] the 2 letter country code to use in the query e.g. 'en'
+ * @property {String|Number} [days] number of days to return in the forecast (used only in dailyForecast alias for cnt)
+ * @property {String|Number} [hours] number of hours to return in the forecast (used only in forecast alias for cnt)
+ */
+
+/**
+ * @typedef {Object} LatLonReqParams
+ * @global
+ * @description an object of parameters used to identify a location using a citys name
+ * @property {Coordinate} coordinates a coordinate object used for searching based on latlon (use on its own)
+ * @property {String|Number} [days] number of days to return in the forecast (used only in dailyForecast alias for cnt)
+ * @property {String|Number} [hours] number of hours to return in the forecast (used only in forecast alias for cnt)
+ */
+
+/**
+ * @typedef {Object} ZipReqParams
+ * @global
+ * @description an object of parameters used to identify a location using a citys name
+ * @property {String|Number} [zip] the zip code to use within the query (US by default unless country is specified)
+ * @property {String|Number} [postcode] an alias to zip
+ * @property {String} [country] the 2 letter country code to use in the query e.g. 'en'
+ * @property {String|Number} [days] number of days to return in the forecast (used only in dailyForecast alias for cnt)
+ * @property {String|Number} [hours] number of hours to return in the forecast (used only in forecast alias for cnt)
+ */
+
+ /**
+ * @typedef {Object} Coordinate
+ * @global
+ * @description a coordinate object containing a longitute and latitude for positioning
+ *
+ * @property {String|Number} latitude the latitude of the position
+ * @property {String|Number} longitude the longitude of the position
+ */
