+++ package.json
@@ -8,7 +8,7 @@
     "openweather",
     "API"
   ],
-  "version": "1.1.0",
+  "version": "1.2.0",
   "author": {
     "name": "Steven Collins",
     "email": "steven@carboncollins.uk"
